requirements.txt
requests==2.31.0
Pillow==9.2.0
ttkthemes==3.2.2
matplotlib==3.7.1 
